refactor: minor clean of dharma.js
